Move tests to composite
